+++ aws_codeseeder/_bundle.py
@@ -25,7 +25,16 @@ from aws_codeseeder import LOGGER, create_output_dir
 
 
 def _is_valid_image_file(file_path: str) -> bool:
-    for word in ("/build/", "/.mypy_cache/", ".egg-info", "__pycache__", "codeseeder.out", "/dist/", "/node_modules/"):
+    for word in (
+        "/build/",
+        "/.mypy_cache/",
+        ".egg-info",
+        "__pycache__",
+        "codeseeder.out",
+        "/dist/",
+        "/node_modules/",
+        "/cdk.out/",
+    ):
         if word in file_path:
             return False
     return True